QCoreApplicationPrivate::initLocale(): report correct encoding
We override the old encoding because it wasn't UTF-8, then we use a
fresh call to nl_langinfo(CODESET) when reporting the encoding that
"is not UTF-8", except that we've just fixed that so it is now.

Store the old encoding in a std::string before we change it, so that
we can report what it was rather than what we changed it to.
Amends commit 3690c202f959a505e0f0bcd4a7b19f235b04d015

Task-number: QTBUG-113371
Change-Id: I5f7c3648890cb0abf1d4769af24715686762c176
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit a7c87be8883458d151e3fe7fa1a8511212af41d4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>