permissions: Remove need to check QT_CONFIG(permissions) in user code
The idea was to avoid having to define return values for platforms
that did not have a permission backend (yet), but the approach does
not work for QML, and was a bit awkward in C++ as well.

We now return Qt::PermissionStatus::Granted on all platforms without
a permission backend, which allows the user code to have a single
code path for checking and requesting permissions. Importantly, that
code path will not be any different now that we always grant the
permission then what it was when the user ifdefed out the permission
check via QT_CONFIG(permissions).

Task-number: QTBUG-90498
Change-Id: I564a24ccfd6b335cb90b7a621778fba61d53a56c
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>