Doc: Correct the expansion of the raisedaster macro for DocBook
QDoc allows code-expanding macros to be defined so that the user can
abstract away common code in a simple to use command.

Furthermore, macros allows for a different expansion based on the
currently generated format so that format-specific requirements can be
satisfied when required.

When format specific expansions are provided, QDoc considers the
resulting text as if it should be generated verbatim.

The "\raisedaster" macro is used to generate an superscript asterisk.

The macro is provided for both HTML and DocBook.
As the macro has format-specific expansions, the expanded code will be
read verbatim.
Nonetheless, while the HTML expansion expands to actual HTML, the
DocBook expansion expands to an equivalent QDoc code, which will not be
generated correctly.

To avoid the issue, the DocBook version of "\raisedaster" is modified so
that it produces the correct DocBook code.

Change-Id: I3a37838bda885af42f8d93b86ec08126d7146ff9
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>