QArrayDataPointer: add missing std::invoke call to projection
To ensure the correct invocation of the projection function, this patch
introduces the usage of std::invoke for all accesses to it. This
modification expands the coverage to include cases where the callable
object is a pointer to a member function.

Amends: 7ca633d9a82f90e5bba5e12ba923bfb0a257af63.

Change-Id: If666012d785ac74c8e856ea9be2a46b3307c8a06
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>